% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_projection.R
\name{matrix_projection}
\alias{matrix_projection}
\title{Matrix Projection of State Transitions}
\usage{
matrix_projection(
  spatial_preprocessing_output = spatial_preprocessing_output,
  spells_analysis_output = spells_analysis_output,
  transition_rules_list = transition_rules_list
)
}
\arguments{
\item{spatial_preprocessing_output}{List output from spatial_preprocessing().}

\item{spells_analysis_output}{List output from spells_analysis().}

\item{transition_rules_list}{Processed transition rules list from format_rules().}
}
\value{
A list containing annual rasters and summarized state transitions.
}
\description{
Projects vegetation state transitions annually using transition matrices and spatial data.
}
