% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rules.R
\name{format_rules}
\alias{format_rules}
\title{Format Transition Rules}
\usage{
format_rules(transition_rules, ordered_states)
}
\arguments{
\item{transition_rules}{Data frame of raw transition rules.}

\item{ordered_states}{Factor of ordered state names.}
}
\value{
A list of processed transition rule components including sequences and spell condition info.
}
\description{
Organizes and augments transition rules to produce spell condition metadata and state transition encoding.
}
