% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Input Data for State Transition Model}
\usage{
prepare_data(
  transition_rules_file,
  flow_files,
  inundation_raster_path,
  vegetation_gdb_path,
  vegetation_layer_name,
  ordered_states
)
}
\arguments{
\item{transition_rules_file}{chr: File path to the CSV containing transition rules.}

\item{flow_files}{list: List of file paths to CSV flow data.}

\item{inundation_raster_path}{chr: File path to the inundation raster.}

\item{vegetation_gdb_path}{chr: Path to the geodatabase with vegetation layer.}

\item{vegetation_layer_name}{chr: Name of the vegetation layer in the GDB.}

\item{ordered_states}{fct: An ordered factor containing ordered states from Highest/First through to Lowest/Worst levels.}
}
\value{
A list containing transition rules, ordered state factors, flow data, inundation raster, and vegetation data.
}
\description{
Loads and processes transition rules, flow data, raster inundation layers, and vegetation layers.
}
